requirements.txt
streamlit
json_repair
langchain_openai
langchain
weaviate-client==3.25.3